split model into several files. Fixed status determination and updates for redis job
