+++ daas/daas_app/models/__init__.py
@@ -1,0 +1,5 @@
+from .redis_job import RedisJob
+from .result import Result
+from .sample import Sample
+from .singals import (report_created_sample_for_statistics, report_sample_result_for_statistics,
+                      revert_sample_result_for_statistics)
