+++ daas/daas_app/models/sample.py
@@ -1,16 +1,12 @@
 from django.db import models
 import hashlib
-import logging
-from django.db.models import Q, Max
+from django.db.models import Q
 from functools import reduce
-from django.db.models.signals import post_save, post_delete
-from django.dispatch import receiver
 
-from .utils.charts import StatisticsManager
-from .utils import redis_status, result_status
-from .utils.redis_manager import RedisManager
-from .config import ALLOW_SAMPLE_DOWNLOAD, SAVE_SAMPLES
-from .utils.configuration_manager import ConfigurationManager
+from ..utils.charts import StatisticsManager
+from ..utils import result_status
+from ..config import ALLOW_SAMPLE_DOWNLOAD, SAVE_SAMPLES
+from ..utils.configuration_manager import ConfigurationManager
 
 
 class SampleQuerySet(models.QuerySet):
@@ -96,11 +92,9 @@ class Sample(models.Model):
         return "%s (type: %s, sha1: %s)" % (self.name, self.file_type, self.sha1)
 
     def status(self):
-        self.redisjob.update()
         return self.redisjob.status
 
     def finished(self):
-        self.redisjob.update()
         return self.redisjob.finished()
 
     def unfinished(self):
@@ -118,10 +112,7 @@ class Sample(models.Model):
 
     @property
     def decompiled(self):
-        try:
-            return self.result.decompiled
-        except AttributeError:
-            return False
+        return self.result.decompiled if self.has_result else False
 
     def content_saved(self):
         return self.data is not None
@@ -169,131 +160,3 @@ class Sample(models.Model):
         if self.has_result:
             StatisticsManager().revert_processed_sample_report(self)
             self.result.delete()
-
-
-class ResultQuerySet(models.QuerySet):
-
-    def failed(self):
-        return self.filter(status=result_status.FAILED)
-
-    def decompiled(self):
-        return self.filter(status=result_status.SUCCESS)
-
-    def timed_out(self):
-        return self.filter(status=result_status.TIMED_OUT)
-
-    def max_elapsed_time(self):
-        max_elapsed_time = self.decompiled().aggregate(Max('elapsed_time'))['elapsed_time__max']
-        return max_elapsed_time if max_elapsed_time is not None else 0
-
-
-class Result(models.Model):
-    class Meta:
-        permissions = (('update_statistics_permission', 'Update Statistics'),)
-
-    timeout = models.SmallIntegerField(default=None, blank=True, null=True)
-    elapsed_time = models.PositiveSmallIntegerField(default=None, blank=True, null=True)
-    exit_status = models.SmallIntegerField(default=None, blank=True, null=True)
-    status = models.PositiveSmallIntegerField(db_index=True)  # fixme: usar choices y charfield
-    output = models.CharField(max_length=10100)
-    compressed_source_code = models.BinaryField(default=None, blank=True, null=True)
-    decompiler = models.CharField(max_length=100)
-    sample = models.OneToOneField(Sample, on_delete=models.CASCADE)
-    processed_on = models.DateTimeField(auto_now_add=True)
-    version = models.SmallIntegerField(default=0)
-    extension = models.CharField(max_length=15)
-
-    objects = ResultQuerySet.as_manager()
-
-    def save(self, *args, **kwargs):
-        # In some strange cases the output was extremely long and higher the limit.
-        if len(self.output) > 10100:
-            logging.debug('Truncating decompiler output. It is too long.')
-            self.output = self.output[:10000] + '\n\n[[[ Output truncated (more than 10000 characters) ]]]'
-        super().save(*args, **kwargs)
-
-    @property
-    def timed_out(self):
-        return self.status == result_status.TIMED_OUT
-
-    @property
-    def failed(self):
-        return self.status == result_status.FAILED
-
-    @property
-    def decompiled(self):
-        return self.status == result_status.SUCCESS
-
-    @property
-    def file_type(self):
-        return self.sample.file_type
-
-    @property
-    def get_config(self):
-        return ConfigurationManager().get_configuration(self.file_type)
-
-    @property
-    def decompiled_with_latest_version(self):
-        return self.version == self.get_config.version
-
-
-class RedisJob(models.Model):
-    class Meta:
-        permissions = (('cancel_job_permission', 'Cancel Job'),)
-
-    job_id = models.CharField(max_length=100)
-    status = models.CharField(default=redis_status.QUEUED, max_length=len(redis_status.PROCESSING))
-    created_on = models.DateTimeField(auto_now_add=True)
-    sample = models.OneToOneField(Sample, on_delete=models.CASCADE)
-
-    def __set_status(self, status):
-        # If we don't use 'save' method here, race conditions might happen, leading to incorrect status.
-        logging.debug('Redis job %s changing status: %s -> %s' % (self.job_id, self.status, status))
-        self.status = status
-        self.save()
-
-    def update(self):
-        if not self.finished():
-            job = RedisManager().get_job(self.sample.file_type, self.job_id)
-            if job is None:
-                self.__set_status(redis_status.DONE if self.sample.decompiled else redis_status.FAILED)
-            elif job.is_finished:
-                self.__set_status(redis_status.DONE)
-            elif job.is_queued:
-                self.__set_status(redis_status.QUEUED)
-            elif job.is_started:
-                self.__set_status(redis_status.PROCESSING)
-            elif job.is_failed:
-                self.__set_status(redis_status.FAILED)
-
-    def finished(self):
-        return self.status in [redis_status.DONE, redis_status.FAILED, redis_status.CANCELLED]
-
-    def is_cancellable(self):
-        return self.status == redis_status.QUEUED
-
-    def is_cancelled(self):
-        return self.status == redis_status.CANCELLED
-
-    def cancel(self):
-        if self.is_cancellable():
-            RedisManager().cancel_job(self.sample.file_type, self.job_id)
-            self.__set_status(redis_status.CANCELLED)
-
-
-# Signals
-@receiver(post_save, sender=Sample)
-def report_created_sample_for_statistics(sender, instance, created, **kwargs):
-    if created:
-        StatisticsManager().report_uploaded_sample(instance)
-
-
-@receiver(post_save, sender=Result)
-def report_sample_result_for_statistics(sender, instance, created, **kwargs):
-    if created:
-        StatisticsManager().report_processed_sample(instance.sample)
-
-
-@receiver(post_delete, sender=Result)
-def revert_sample_result_for_statistics(sender, instance, **kwargs):
-    StatisticsManager().revert_processed_sample_report(instance.sample)
