+++ daas/daas_app/templates/registration/login.html
@@ -1,4 +1,4 @@
-{% extends "daas_app/base.html" %}
+{% extends "daas_app/layout/base.html" %}
 
 {% block title %}Login{% endblock %}
